Ensure Chosen are added to more than one mission (#54)
Had to replicate the code that adds the Chosen tactical tags to mission sites
in `XCGS_LWAlienActivity` so that Chosen are added to missions beyond the
first encounter.

This implementation is very simple and simply rolls for a 25% chance to add
a Chosen to a mission. It also allows the mod to configure mission types that
should not have Chosen under any circumstances. Only Jailbreak_LW is currently
excluded because its timer and reinforcements don't pause when the Chosen
engage XCOM.